Add files via upload
This code integrates ESP32 CAM, Ultrasonic Sensor, IR obstacle detection sensor and HF Mini player with the Arduino Uno. The code communicates with ESP32 CAM over Serial communication which outputs the prediction of the object the ESP32 CAM detected. Based upon the object detected, the Arduino Uno sends the corresponding sound (eg. Wall) to play to the HF Mini player. The Ultrasonic and IR sensors are also working side by side to play the buzzing sound when the object is very close to the blind person.